# api/internal/handler/openapi.gen.json
{
...
  "info": {
    "description": "A modern, agent-friendly frontend for the Python Package Index. Every JSON endpoint listed here that has a `.txt` sibling also honours `Accept: text/plain`, returning the plain-text representation instead. All responses are cacheable; requests are rate limited to 30 req/s (burst 60) per client at the origin, reported via the X-RateLimit-* headers and a 429 response. The edge additionally enforces 60 req/min per IP on /api/*, with a 10-minute block for IPs that exceed it.",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "pypx API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/openapi.json": {
      "get": {
        "description": "The generated OpenAPI 3.1 spec describing this JSON API, including this route. This route has no plain-text twin; `Accept: text/plain` is ignored. Server-side timeout: 30s.",
        "operationId": "getApiOpenapiJson",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Success."
          },
          "429": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Rate limit exceeded. See the X-RateLimit-* and Retry-After headers.",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              },
              "X-RateLimit-Limit": {
                "description": "Configured requests-per-second rate, as a string.",
                "schema": {
                  "type": "string"
                }
              },
              "X-RateLimit-Remaining": {
                "description": "Remaining requests in the current burst window.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "default": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Error. Upstream failures surface as 502; internal failures as 500."
          }
        },
        "summary": "OpenAPI 3.1 document"
      }
    },
...
    "/llms.txt": {
      "get": {
        "description": "Index of the plain-text and JSON endpoints available on this API, per the llms.txt convention. Server-side timeout: 30s.",
        "operationId": "getLlmsTxt",
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Plain-text representation."
          },
          "429": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Rate limit exceeded. See the X-RateLimit-* and Retry-After headers.",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              },
              "X-RateLimit-Limit": {
                "description": "Configured requests-per-second rate, as a string.",
                "schema": {
                  "type": "string"
                }
              },
              "X-RateLimit-Remaining": {
                "description": "Remaining requests in the current burst window.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "default": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Error. Upstream failures surface as 502; internal failures as 500."
          }
        },
        "summary": "Agent discovery index (text only)"
      }
    }
  },
  "servers": [
    {
      "description": "Production",
      "url": "https://pypx.app"
    }
  ]
}